requirements.txt
google-genai>=0.1.0
opencv-python>=4.8.0
pillow>=10.0.0
python-dotenv>=1.0.0
keyboard>=0.13.5
gTTS>=2.4.0
pygame>=2.5.0
RPi.GPIO>=0.7.1 